requirements.txt
streamlit
feedparser
requests
pyyaml
langchain
langchain-groq
langchain-core
pandas
plotly
python-dotenv
pydub
psycopg2-binary
sqlalchemy
